Add files via upload
Labeled Datasets: The activities in the generated datasets were labeled as follows: vision-reaching (1), no-vision reaching (2), vision grasping (3), and no-vision grasping (4). These labels were added as an additional column at the end of the dataset. For analytical purposes, we removed the likelihood column since it was not relevant to our analysis. Additionally, we merged the first two header rows into one to simplify the datasets for easier analysis later.